+++ test/integration/run.sh
@@ -4,12 +4,13 @@ set -e
 
 export CLICKHOUSE_VERSION=${CLICKHOUSE_VERSION:-21.8}
 export CLICKHOUSE_IMAGE=${CLICKHOUSE_IMAGE:-yandex/clickhouse-server}
-export CLICKHOUSE_BACKUP_BIN="$(pwd)/clickhouse-backup/clickhouse-backup"
+export CLICKHOUSE_BACKUP_BIN="$(pwd)/clickhouse-backup/clickhouse-backup-race"
 export LOG_LEVEL=${LOG_LEVEL:-info}
 export GCS_TESTS=${GCS_TESTS:-}
 export AZURE_TESTS=${AZURE_TESTS:-}
 export S3_DEBUG=${S3_DEBUG:-false}
 export GCS_DEBUG=${GCS_DEBUG:-false}
+export FTP_DEBUG=${FTP_DEBUG:-false}
 export GODEBUG=${GODEBUG:-}
 
 if [[ "${CLICKHOUSE_VERSION}" == 2* ]]; then
@@ -21,6 +22,6 @@ fi
 docker-compose -f test/integration/${COMPOSE_FILE} down --remove-orphans
 docker volume prune -f
 make clean
-make build
+make build-race
 docker-compose -f test/integration/${COMPOSE_FILE} up -d --force-recreate
 go test  -timeout 30m -failfast -tags=integration -run "${RUN_TESTS:-.+}" -v test/integration/integration_test.go